Move positioning logic to CanvasPanel
