+++ bokehjs/src/lib/models/canvas/canvas_panel.ts
@@ -1,0 +1,93 @@
+import {StyledElement, StyledElementView} from "../ui/styled_element"
+import type {PlotView} from "../plots/plot"
+import type {Node} from "../coordinates/node"
+import type * as p from "core/properties"
+import {InlineStyleSheet} from "core/dom"
+import type {StyleSheetLike} from "core/dom"
+import type {XY} from "core/util/bbox"
+import {BBox} from "core/util/bbox"
+import {isNumber} from "core/util/types"
+
+export class CanvasPanelView extends StyledElementView {
+  declare model: CanvasPanel
+  declare parent: PlotView
+
+  private _bbox: BBox = new BBox()
+  override get bbox(): BBox {
+    return this._bbox
+  }
+
+  readonly position = new InlineStyleSheet()
+
+  override stylesheets(): StyleSheetLike[] {
+    return [...super.stylesheets(), this.position]
+  }
+
+  override rendering_target(): HTMLElement {
+    return this.parent.canvas_view.events_el
+  }
+
+  set_geometry(bbox: BBox): void {
+    this._bbox = bbox
+    this._update_position()
+  }
+
+  /**
+   * Updates the position of the associated DOM element.
+   */
+  protected _update_position(): void {
+    const {bbox, position} = this
+    if (bbox.is_valid) {
+      position.replace(`
+      :host {
+        position: absolute;
+        left:     ${bbox.left}px;
+        top:      ${bbox.top}px;
+        width:    ${bbox.width}px;
+        height:   ${bbox.height}px;
+        overflow: hidden;
+        pointer-events: none;
+      }
+      `)
+    } else {
+      position.replace(`
+      :host {
+        display: none;
+      }
+      `)
+    }
+  }
+
+  // TODO remove this when bbox handling is unified
+  override resolve_symbol(node: Node): XY | number {
+    const target = this
+    const value = target.bbox.resolve(node.symbol)
+    const {offset} = node
+    if (isNumber(value)) {
+      return value + offset
+    } else {
+      const {x, y} = value
+      return {x: x + offset, y: y + offset}
+    }
+  }
+}
+
+export namespace CanvasPanel {
+  export type Attrs = p.AttrsOf<Props>
+  export type Props = StyledElement.Props
+}
+
+export interface CanvasPanel extends CanvasPanel.Attrs {}
+
+export class CanvasPanel extends StyledElement {
+  declare properties: CanvasPanel.Props
+  declare __view_type__: CanvasPanelView
+
+  constructor(attrs?: Partial<CanvasPanel.Attrs>) {
+    super(attrs)
+  }
+
+  static {
+    this.prototype.default_view = CanvasPanelView
+  }
+}
