+++ bokehjs/src/lib/models/canvas/cartesian_frame.ts
@@ -1,5 +1,4 @@
-import {StyledElement, StyledElementView} from "../ui/styled_element"
-import type {PlotView} from "../plots/plot"
+import {CanvasPanel, CanvasPanelView} from "./canvas_panel"
 import {CategoricalScale} from "../scales/categorical_scale"
 import {LogScale} from "../scales/log_scale"
 import {Scale} from "../scales/scale"
@@ -8,28 +7,17 @@ import {Range} from "../ranges/range"
 import {Range1d} from "../ranges/range1d"
 import {DataRange1d} from "../ranges/data_range1d"
 import {FactorRange} from "../ranges/factor_range"
-import type {Node} from "../coordinates/node"
-import type {XY} from "core/util/bbox"
-import {BBox} from "core/util/bbox"
+import type {BBox} from "core/util/bbox"
 import {entries} from "core/util/object"
 import {assert} from "core/util/assert"
-import {isNumber} from "core/util/types"
 import type {Dict} from "core/types"
 import type * as p from "core/properties"
-import {InlineStyleSheet} from "core/dom"
-import type {StyleSheetLike} from "core/dom"
 
 type Ranges = Dict<Range>
 type Scales = Dict<Scale>
 
-export class CartesianFrameView extends StyledElementView {
+export class CartesianFrameView extends CanvasPanelView {
   declare model: CartesianFrame
-  declare parent: PlotView
-
-  private _bbox: BBox = new BBox()
-  override get bbox(): BBox {
-    return this._bbox
-  }
 
   override initialize(): void {
     super.initialize()
@@ -144,10 +132,9 @@ export class CartesianFrameView extends StyledElementView {
     }
   }
 
-  set_geometry(bbox: BBox): void {
-    this._bbox = bbox
+  override set_geometry(bbox: BBox): void {
+    super.set_geometry(bbox)
     this._update_scales()
-    this._update_position()
   }
 
   get x_range(): Range {
@@ -197,62 +184,12 @@ export class CartesianFrameView extends StyledElementView {
   get y_scale(): Scale {
     return this._y_scale
   }
-
-  // TODO remove this when bbox handling is unified
-  override resolve_symbol(node: Node): XY | number {
-    const target = this
-    const value = target.bbox.resolve(node.symbol)
-    const {offset} = node
-    if (isNumber(value)) {
-      return value + offset
-    } else {
-      const {x, y} = value
-      return {x: x + offset, y: y + offset}
-    }
-  }
-
-  readonly position = new InlineStyleSheet()
-
-  override stylesheets(): StyleSheetLike[] {
-    return [...super.stylesheets(), this.position]
-  }
-
-  override rendering_target(): HTMLElement {
-    return this.parent.canvas_view.events_el
-  }
-
-  /**
-   * Updates the position of the associated DOM element.
-   */
-  protected _update_position(): void {
-    const {bbox, position} = this
-    if (bbox.is_valid) {
-      position.replace(`
-      :host {
-        position: absolute;
-        left:     ${bbox.left}px;
-        top:      ${bbox.top}px;
-        width:    ${bbox.width}px;
-        height:   ${bbox.height}px;
-        overflow: hidden;
-        pointer-events: none;
-      }
-      `)
-    } else {
-      position.replace(`
-      :host {
-        display: none;
-      }
-      `)
-    }
-  }
-
 }
 
 export namespace CartesianFrame {
   export type Attrs = p.AttrsOf<Props>
 
-  export type Props = StyledElement.Props & {
+  export type Props = CanvasPanel.Props & {
     x_range: p.Property<Range>
     y_range: p.Property<Range>
 
@@ -272,7 +209,7 @@ export namespace CartesianFrame {
 
 export interface CartesianFrame extends CartesianFrame.Attrs {}
 
-export class CartesianFrame extends StyledElement {
+export class CartesianFrame extends CanvasPanel {
   declare properties: CartesianFrame.Props
   declare __view_type__: CartesianFrameView
